Move pkg bound to bounds.disco
Later, there should be interfaces in the bounds package to allow for
compatible new implementations
