+++ src/main/java/de/uni_kl/cs/discodnc/bounds/disco/Output.java
@@ -26,7 +26,7 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.bounds;
+package de.uni_kl.cs.discodnc.bounds.disco;
 
 import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
