+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/analyses/TandemMatchingAnalysis.java
@@ -39,7 +39,7 @@ import java.util.Map.Entry;
 import org.apache.commons.math3.util.Pair;
 
 import de.uni_kl.cs.discodnc.Calculator;
-import de.uni_kl.cs.discodnc.bounds.Bound;
+import de.uni_kl.cs.discodnc.bounds.disco.Bound;
 import de.uni_kl.cs.discodnc.numbers.Num;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
