+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/arrivalbounds/AggregatePmoo.java
@@ -29,7 +29,7 @@
 
 package de.uni_kl.cs.discodnc.feedforward.arrivalbounds;
 
-import de.uni_kl.cs.discodnc.bounds.Bound;
+import de.uni_kl.cs.discodnc.bounds.disco.Bound;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
